requirements.txt
Django==3.1
six==1.15.0
djangorestframework==3.11.1
dj-rest-auth==1.1.2
django-cors-headers==3.4.0
django-allauth==0.43.0
django-extensions==3.0.5
djangorestframework-api-key==2.0.0
django-oauth-toolkit==1.3.2
django-storages==1.10.1
psycopg2==2.8.6
gunicorn==20.0.4
requests==2.24.0
celery==4.4.7
redis==3.5.3
ipython==7.19.0
google-api-core==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-cloud-logging==1.15.1
google-auth==1.23.0
factory-boy==3.1.0
coverage==5.3
stripe==2.55.0